Refactor DeliverySystemCLI and PositiveIntegerValidator for improved type inference and structure

+++ src/main/java/services/OrderService.java
@@ -6,7 +6,7 @@ import model.MenuItem;
 import model.Order;
 
 public interface OrderService {
-    Order getOrderById(Long orderId);
+    Order getOrderById(Integer orderId);
     Order createOrder(List<MenuItem> items);
 
     void displayOrderDetails(Order order);
